+++ vars/sectionDeployToAKS.groovy
@@ -32,12 +32,16 @@ def call(params) {
   Builder builder = pipelineType.builder
 
   if (config.dockerBuild) {
-    withAksClient(subscription, environment) {
-      def acr = new Acr(this, subscription, env.REGISTRY_NAME, env.REGISTRY_RESOURCE_GROUP)
-      def dockerImage = new DockerImage(product, component, acr, new ProjectBranch(env.BRANCH_NAME).imageTag(), env.GIT_COMMIT)
+    def acr = new Acr(this, subscription, env.REGISTRY_NAME, env.REGISTRY_RESOURCE_GROUP)
+    def dockerImage = new DockerImage(product, component, acr, new ProjectBranch(env.BRANCH_NAME).imageTag(), env.GIT_COMMIT)
+
+    withAcrClient(subscription) {
       onPR {
         acr.retagForStage(DockerImage.DeploymentStage.PR, dockerImage)
       }
+    }
+    withAksClient(subscription, environment) {
+
       if (config.deployToAKS) {
 
         withTeamSecrets(config, environment) {
