+++ vars/sectionPromoteBuildToStage.groovy
@@ -37,7 +37,7 @@ def call(params) {
 
   stage("${deploymentStage.label} build promotion") {
     if (config.dockerBuild) {
-      withAksClient(subscription, params.environment) {
+      withAcrClient(subscription) {
 
         def acr = new Acr(this, subscription, env.REGISTRY_NAME, env.REGISTRY_RESOURCE_GROUP)
         def dockerImage = new DockerImage(product, component, acr, new ProjectBranch(env.BRANCH_NAME).imageTag(), env.GIT_COMMIT)
